Change Ingress controller to common
